# docs/hld/openapi.yaml
...
openapi: 3.1.0
info:
  title: System-112 / DDS training simulator — backend API
  version: 1.0.0
  summary: Commands and resources for the local emergency-response training simulator.
  description: |
    REST command/resource API under `/api/v1` (D8). Realtime session updates travel over the
    WebSocket channel specified in `docs/hld/40-realtime-protocol.md`; raw microphone audio
    travels over LiveKit and never over this API (SPEC §34).

    Authorisation of every command is two-gated (D8): the caller's participant record must be
    assigned to the session's active `RoleStage`, **and** the stage's `RoleModule` must permit
    the action in the current stage state. A failure of the first gate is `403 FORBIDDEN_FOR_ROLE`;
    of the second, `409 ACTION_NOT_AVAILABLE`. An illegal state-machine trigger is
    `409 INVALID_TRANSITION`.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local docker-compose backend (SPEC §36).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [sessions]
      operationId: createSession
      summary: Create a simulation session (INSTRUCTOR / ADMIN).
      description: |
        Creates one `SimulationSession`, one `Incident` and one `RoleStage` per `role_chain`
        entry, instantiates `WorldTruth` and `CallerBelief` from the version, then fires the
        session machine trigger `validate` (`CREATED → READY`). A `role_chain` of `[DDS]` under
        `SINGLE_ROLE` or `ASSESSMENT` without `expected_response.prefab_handoff` is refused with
        `409 PREFAB_HANDOFF_REQUIRED` (D6, §10.10).
      x-action: '-'
      x-emits: [SESSION_CREATED, ROLE_STAGE_STARTED]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCreateRequest' }
      responses:
        '201':
          description: The created session in state `READY` (or `CREATED` if validation deferred).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [sessions]
      operationId: listSessions
      summary: List sessions — mine, or all for an instructor.
      description: |
        `scope=MINE` returns the sessions the caller participates in or created.
        `scope=ALL` is permitted for `INSTRUCTOR` and `ADMIN` only; a `TRAINEE` asking for
        `ALL` gets `403 FORBIDDEN_FOR_ROLE`.
      x-action: '-'
      x-emits: []
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [MINE, ALL]
            default: MINE
        - name: state
          in: query
          required: false
          description: Filter by session state.
          schema: { $ref: '#/components/schemas/SessionState' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Sessions, newest first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SessionListItem' }
                  total: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/sessions/{session_id}/abort:
    post:
      tags: [sessions]
      operationId: abortSession
      summary: Abort the session (INSTRUCTOR / ADMIN).
      description: |
        Fires `abort` on the session machine and `abort_stage` on the active `RoleStage`.
        Simulation data is preserved: the event log is closed, never deleted (SPEC §39,
        §42 test 14).
      x-action: abort
      x-emits: [SESSION_ABORTED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AbortSessionRequest' }
      responses:
        '200':
          description: The session in state `ABORTED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/sessions/{session_id}/operator/call/answer:
    post:
      tags: [operator]
      operationId: answerCall
      summary: Answer the ringing call.
      description: |
        Fires `answer` (`RINGING → CONNECTED`, guard `guard_participant_assigned_to_stage`).
        Permission `ANSWER_CALL`. The frontend then joins the LiveKit room with the token from
        `createVoiceToken`; the voice-agent is already in the room, having joined on the
        `voice:join` message published at `CALL_RINGING` (D9).
      x-action: answer
      x-emits: [CALL_ANSWERED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new operator stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/operator/call/end:
    post:
      tags: [operator]
      operationId: endCall
      summary: Hang up the call.
      description: |
        Permission `END_CALL`. Not a stage transition of its own: it ends the call and is the
        precondition of `complete_stage` (guard `guard_call_ended`). The backend publishes
        `voice:cancel:{session_id}` so an in-flight caller response is cancelled across
        processes (D9).
      x-action: end_call
      x-emits: [CALL_ENDED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EndCallRequest' }
      responses:
        '200':
          description: The new operator stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/operator/card:
    get:
      tags: [operator]
      operationId: getOperatorCard
      summary: The current incident card.
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The card.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorCardView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{session_id}/operator/card/field:
    put:
      tags: [operator]
      operationId: setCardField
      summary: Set exactly one card field.
      description: |
        Permission `EDIT_CARD`. Exactly one `field_path` per command — the API has no bulk card
        write, because SPEC §9 requires every mutation to be stored with its previous value and
        actor, and a bulk write would blur which change happened when. The trainee is the only
        writer: ASR never calls this (SPEC §9, §42 test 4), which is structural — the voice-agent
        has no credential for it.

        `field_path` must be a member of `CARD_FIELDS` (`10-domain-model.md` §10.6) and the value
        must match that field's `value_type`; otherwise `422 CARD_FIELD_UNKNOWN` or
        `422 CARD_VALUE_TYPE_MISMATCH`. Setting a field to its current value is a no-op: it
        returns `revision: null`, appends no revision and emits no event.

        `recipients.services` is **not** settable here; use the service commands.
      x-action: edit_card
      x-emits: [CARD_FIELD_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetCardFieldRequest' }
      responses:
        '200':
          description: The new card plus the `CardRevision` this command created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetCardFieldResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/sessions/{session_id}/operator/card/revisions:
    get:
      tags: [operator]
      operationId: listCardRevisions
      summary: The card's revision history.
      description: |
        Source `CARD_REVISIONS` of the operator's `DataVisibilityPolicy`. This is what makes
        "the final card is NOT enough" (SPEC §9) visible in the UI and the report.
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: field_path
          in: query
          required: false
          description: Optional filter to one card field path.
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Revisions ordered by ascending `revision_no`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardRevisionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{session_id}/operator/services/select:
    post:
      tags: [operator]
      operationId: selectRecipientService
      summary: Add one recipient service to the card.
      description: |
        Permission `SELECT_SERVICES`. Writes `recipients.services`, so it emits
        `CARD_FIELD_CHANGED` **in addition to** `SERVICE_SELECTED`: both the service evaluators
        and the card evaluators must be self-sufficient from the log alone (D5,
        `10-domain-model.md` §10.6). Selecting an already selected service is a no-op.
      x-action: select_services
      x-emits: [CARD_FIELD_CHANGED, SERVICE_SELECTED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceSelectionRequest' }
      responses:
        '200':
          description: The new recipient-service selection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceSelectionView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/sessions/{session_id}/operator/services/deselect:
    post:
      tags: [operator]
      operationId: deselectRecipientService
      summary: Remove one recipient service from the card.
      x-action: select_services
      x-emits: [CARD_FIELD_CHANGED, SERVICE_DESELECTED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceSelectionRequest' }
      responses:
        '200':
          description: The new recipient-service selection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceSelectionView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/sessions/{session_id}/operator/handoff/prepare:
    post:
      tags: [operator]
      operationId: beginHandoffPreparation
      summary: Open the handoff preparation screen.
      description: |
        Fires `open_handoff_preparation` (`INTERVIEW → HANDOFF_PREPARATION`). Deliberately
        unguarded: an incomplete card must remain possible, because an omission must propagate
        to DDS as a real mistake (SPEC §10).
      x-action: open_handoff_preparation
      x-emits: [STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new operator stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/operator/handoff/cancel:
    post:
      tags: [operator]
      operationId: backToInterview
      summary: Return from handoff preparation to the interview.
      description: |
        Fires `back_to_interview` (`HANDOFF_PREPARATION → INTERVIEW`, guard
        `guard_call_still_connected`).
      x-action: back_to_interview
      x-emits: [STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new operator stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/operator/handoff:
    post:
      tags: [operator]
      operationId: createHandoff
      summary: Freeze the card and send it to DDS.
      description: |
        Permission `CREATE_HANDOFF`; fires `create_handoff`
        (`HANDOFF_PREPARATION → HANDED_OFF`, guard `guard_at_least_one_recipient_service` —
        otherwise `409 RECIPIENT_SERVICES_EMPTY`).

        In one Unit of Work: the optional `comment_ru` is written to the card field
        `recipients.comment` first (a normal revision, so a `CARD_FIELD_CHANGED` precedes the
        three events below), `freeze_card_to_snapshot` deep-copies `OperatorCard.values` into an
        immutable `HandoffSnapshot`, and `snapshot_to_assignments` creates one `DDSAssignment`
        per recipient service. `HANDOFF_CREATED` is emitted exactly once whatever the number of
        recipients — one trainee action, one trainee event — and `HANDOFF_RECEIVED` once per
        assignment, `SIMULATION`-authored; those are the log's only record of which assignment
        belongs to which service. Nothing is read from `WorldTruth`
        (SPEC §10, §42 test 3): if the operator entered house `72` where the world says `27`,
        DDS receives `72`.
      x-action: create_handoff
      x-emits: [HANDOFF_CREATED, STAGE_STATE_CHANGED, HANDOFF_RECEIVED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHandoffRequest' }
      responses:
        '201':
          description: The frozen snapshot and the assignments it created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HandoffCreatedView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/operator/stage/complete:
    post:
      tags: [operator]
      operationId: completeOperatorStage
      summary: Complete the Operator 112 stage.
      description: |
        Fires `complete_stage` (`HANDED_OFF → STAGE_COMPLETED`, guard `guard_call_ended`).
        The session machine then fires `begin_role_transition` when the `role_chain` has a next
        entry, or `complete` when it does not.
      x-action: complete_stage
      x-emits: [ROLE_STAGE_COMPLETED, STAGE_STATE_CHANGED, ROLE_TRANSITION_STARTED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The session after stage completion.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ------------------------------------------------------ role transition ----
  /api/v1/sessions/{session_id}/stage/continue:
    post:
      tags: [sessions]
      operationId: continueToNextStage
      summary: Continue to the next role stage after the configurable pause.
      description: |
        Fires `finish_role_transition` (`ROLE_TRANSITION → ACTIVE`, guard
        `guard_pause_elapsed_and_next_assigned`: `now_ms ≥ transition_started_ms +
        policy.transition_pause_seconds·1000`). Called before the pause elapses it is refused
        with `409 INVALID_TRANSITION` — the pause is `SessionPolicy.transition_pause_seconds`
        (§10.10), not a frontend timer.

        The same `Incident` carries over; no new incident is created (SPEC §13, §42 test 5).
      x-action: finish_role_transition
      x-emits: [ROLE_TRANSITION_COMPLETED, ROLE_STAGE_STARTED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The session with the next stage active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # -------------------------------------------------- DDS stage commands ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT minted by `loginUser`; the secret comes from the environment (D8, SPEC §41).
...
  responses:
    Unauthorized:
      description: '`UNAUTHENTICATED` — missing, malformed or expired bearer token.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: |
        `FORBIDDEN_FOR_ROLE` — the caller's user role, or the participant assignment to the
        active `RoleStage`, does not permit this operation. Also `REPORT_NOT_RELEASED`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: '`NOT_FOUND` — no such session, scenario, version, resource or segment.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    RoleType:
      type: string
      enum: [OPERATOR_112, DDS, EDDS]
...
    SessionState:
      type: string
      enum: [CREATED, READY, ACTIVE, ROLE_TRANSITION, COMPLETED, ABORTED]
...
    SessionCreateRequest:
      type: object
      additionalProperties: false
      required: [scenario_version_id, session_mode, participants]
      properties:
        scenario_version_id: { type: string, format: uuid }
        session_mode: { $ref: '#/components/schemas/SessionMode' }
        participants:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ParticipantAssignment' }
        session_seed:
          type: [string, 'null']
          description: |
            Overrides `ScenarioVersion.deterministic_seed`. Recorded in the `SESSION_CREATED`
            payload, so an identical seed with identical timestamped actions reproduces an
            identical world-event stream (D7, SPEC §42 test 7).
        time_scale:
          type: number
          minimum: 0.1
          maximum: 10
          default: 1
          description: |
            Multiplier applied to simulated time only. It scales world-event scheduling and ETA
            countdowns; it never scales the audio path, which is realtime by nature.
...
    SessionDetail:
      type: object
      description: |
        The materialized session header every command returns (D8). `session_seed` is included
        because it is public reproduction data, not hidden incident truth.
      additionalProperties: false
      required:
        - id
        - scenario_version_id
        - scenario_slug
        - scenario_version
        - session_mode
        - state
        - session_seed
        - time_scale
        - incident_id
        - role_chain
        - stages
        - active_role_stage_id
        - participants
        - created_by_user_id
        - created_at
        - started_at
        - completed_at
        - abort_reason
        - monotonic_offset_ms
        - last_seq_no
        - transition_pause_seconds
        - transition_continue_available_at_offset_ms
      properties:
        id: { type: string, format: uuid }
        scenario_version_id: { type: string, format: uuid }
        scenario_slug: { type: string }
        scenario_version: { type: integer, minimum: 1 }
        session_mode: { $ref: '#/components/schemas/SessionMode' }
        state: { $ref: '#/components/schemas/SessionState' }
        session_seed: { type: string }
        time_scale: { type: number }
        incident_id: { type: string, format: uuid }
        role_chain:
          type: array
          items: { $ref: '#/components/schemas/RoleType' }
        stages:
          type: array
          items: { $ref: '#/components/schemas/RoleStageView' }
        active_role_stage_id: { type: [string, 'null'], format: uuid }
        participants:
          type: array
          items: { $ref: '#/components/schemas/SessionParticipantView' }
        created_by_user_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        abort_reason: { type: [string, 'null'] }
        monotonic_offset_ms:
          type: integer
          minimum: 0
          description: |
            Current simulation time: ms since `SESSION_STARTED`, derived from the persisted
            `started_at` plus paused intervals, so a restart or a refresh never resets it (D7).
        last_seq_no:
          type: integer
          minimum: 0
          description: The highest `seq_no` visible to the caller's role — the WebSocket resume cursor.
        transition_pause_seconds:
          type: integer
          minimum: 0
          description: '`SessionPolicy.transition_pause_seconds` for this mode (§10.10).'
        transition_continue_available_at_offset_ms:
          type: [integer, 'null']
          description: |
            When `state` is `ROLE_TRANSITION`, the offset at which `continueToNextStage` stops
            being refused. `null` otherwise.
...